CfpRaDecValidator
